+++ tree/phylotreemixlen.cpp
@@ -829,7 +829,7 @@ void PhyloTreeMixlen::computeFuncDerv(double value, double &df, double &ddf) {
     for (size_t c = 0; c < ncat; c++) {
         double prop = site_rate->getProp(c);
         for (size_t i = 0; i < nstates; i++) {
-            double cof = eval[cur_mixture*nstates+i]*site_rate->getRate(c);
+            double cof = eval[cur_mixture*nstates+i]*site_rate->getRate(static_cast<int>(c));
             // length for heterotachy model
             double val = exp(cof*dad_branch->getLength(cur_mixture)) * prop * model->getMixtureWeight(cur_mixture);
             double val1_ = cof*val;
